# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the backend server for the Thundermeet App.",
        "title": "ThunderMeet APIs",
        "contact": {
            "name": "Wu, Chien Yin and Yeh, Hsiao Li"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
        "/v1/users": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "parameters": [
                    {
                        "description": "The body to create a user",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Register"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 31a165baebe6dec616b1f8f3207b4273",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "The body to create a user",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Update"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/users/": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 31a165baebe6dec616b1f8f3207b4273",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/users/login": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "parameters": [
                    {
                        "description": "The body to login a user",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Login"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/users/resetPassword": {
            "patch": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "parameters": [
                    {
                        "description": "The body to create a user",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.ForgotInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "Login": {
            "type": "object",
            "required": [
                "password",
                "userId"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "example": "password"
                },
                "userId": {
                    "type": "string",
                    "example": "christine891225"
                }
            }
        },
        "Register": {
            "type": "object",
            "required": [
                "password",
                "passwordAnswer",
                "userId"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "example": "password"
                },
                "passwordAnswer": {
                    "type": "string",
                    "example": "NTU"
                },
                "userId": {
                    "type": "string",
                    "example": "christine891225"
                },
                "userName": {
                    "type": "string",
                    "example": "Christine Wang"
                }
            }
        },
        "Update": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string",
                    "example": "password"
                },
                "passwordAnswer": {
                    "type": "string",
                    "example": "NTU"
                },
                "userName": {
                    "type": "string",
                    "example": "Christine Wang"
                }
            }
        },
        "controller.ForgotInfo": {
            "type": "object",
            "required": [
                "password",
                "passwordAnswer",
                "userId"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "example": "password"
                },
                "passwordAnswer": {
                    "type": "string",
                    "example": "NTU"
                },
                "userId": {
                    "type": "string",
                    "example": "christine891225"
                }
            }
        }
    }
}